+++ src/popups/StagesPopup/UpdateScrollLayoutEvent.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include <Geode/Geode.hpp>
+#include <Geode/loader/Event.hpp>
+
+using namespace geode::prelude;
+
+class UpdateScrollLayoutEvent : public Event
+{
+};