+++ src/popups/StagesPopup/ui/StageList/StageRangeCell.cpp
@@ -32,22 +32,46 @@ bool StageRangeCell::init(Range *range, GJGameLevel *level, const CCSize &cellSi
       m_isCurrent = currentRange.id == m_range->id;
   }
 
+  // ! --- Content --- !
+  m_content = CCLayer::create();
+  m_content->setZOrder(1);
+  m_content->setContentSize(cellSize);
+  m_content->setPosition({0, cellSize.height});
+  m_content->setAnchorPoint({0, 1});
+  m_content->setLayout(
+      ColumnLayout::create()
+          ->setGap(0)
+          ->setAxisReverse(true)
+          ->setAutoScale(false)
+          ->setAutoGrowAxis(true));
+  m_content->getLayout()->ignoreInvisibleChildren(true);
+  m_content->setID("progress-cell-content"_spr);
+  this->addChild(m_content);
+  updateLayoutWrapper();
+
+  // ! --- Head --- !
+  m_head = CCLayer::create();
+  m_head->setContentSize(cellSize);
+  m_head->setID("progress-cell-head"_spr);
+  m_content->addChild(m_head);
+  updateLayoutWrapper();
+
+  // ! --- Background --- !
   updateBackgroundTexture();
 
   // ! --- Range Label --- !
   const float diff = std::abs(m_range->from - m_range->to);
 
-  m_rangeLabel = Label::create(
-      fmt::format(
-          "{:.0f}<small>.{:02.0f}%</small> - {:.0f}<small>.{:02.0f}%</small>",
-          std::floor(m_range->from),
-          std::round((m_range->from - std::floor(m_range->from)) * 100),
-          std::floor(m_range->to),
-          std::round((m_range->to - std::floor(m_range->to)) * 100)));
+  std::string rangeText = fmt::format(
+      "{:.0f}<small>.{:02.0f}%</small> - {:.0f}<small>.{:02.0f}%</small>",
+      std::floor(m_range->from),
+      std::round((m_range->from - std::floor(m_range->from)) * 100),
+      std::floor(m_range->to),
+      std::round((m_range->to - std::floor(m_range->to)) * 100));
+  m_rangeLabel = Label::create(rangeText, "gjFont17.fnt", .4f);
   m_rangeLabel->setPosition({25.f, cellSize.height / 2});
   m_rangeLabel->setAnchorPoint({0.f, .5f});
-  m_rangeLabel->setZOrder(1);
-  this->addChild(m_rangeLabel);
+  m_head->addChild(m_rangeLabel);
   updateTextColors();
 
   // ! --- Checkbox --- !
@@ -57,24 +81,74 @@ bool StageRangeCell::init(Range *range, GJGameLevel *level, const CCSize &cellSi
   m_checkbox = CCMenuItemToggler::create(toggleOn, toggleOff, this, menu_selector(StageRangeCell::onToggle));
   m_checkbox->setAnchorPoint({0.5f, 0.5f});
   m_checkbox->setCascadeColorEnabled(true);
-  m_checkbox->setZOrder(1);
   m_checkbox->toggle(!m_checked);
 
-  // ! --- Menu --- !
-  auto menu = CCMenu::createWithItem(m_checkbox);
-  menu->setScale(0.5f);
-  menu->setAnchorPoint({0.5f, 0.f});
-  menu->setContentSize({32.f, 32.f});
-  menu->setPosition({5.f, cellSize.height / 2});
-  menu->setZOrder(1);
-  this->addChild(menu);
+  // ! --- Checkbox Menu --- !
+  auto checkboxMenu = CCMenu::createWithItem(m_checkbox);
+  checkboxMenu->setScale(0.5f);
+  checkboxMenu->setAnchorPoint({0.5f, 0.f});
+  checkboxMenu->setContentSize({32.f, 32.f});
+  checkboxMenu->setPosition({5.f, cellSize.height / 2});
+  m_head->addChild(checkboxMenu);
+
+  // ! --- Expand Button --- !
+  updateExpandButton();
+
+  // ! --- Meta Info Content --- !
+  bool isFirstRunExists = m_range->firstRunFrom >= 0 && m_range->firstRunTo > 0;
+  bool isBestRunExists = m_range->bestRunFrom >= 0 && m_range->bestRunTo > 0;
+
+  std::string attempts = fmt::format("{}", m_range->attempts);
+  std::string completions = fmt::format("{}<small>times</small>", m_range->completionCounter);
+  std::string firstRun = !isFirstRunExists
+                             ? "<small>None</small>"
+                             : fmt::format(
+                                   "{}<small>.{:02d}%</small> - {}<small>.{:02d}%</small>",
+                                   static_cast<int>(m_range->firstRunFrom),
+                                   static_cast<int>(std::round((m_range->firstRunFrom - static_cast<int>(m_range->firstRunFrom)) * 100)),
+                                   static_cast<int>(m_range->firstRunTo),
+                                   static_cast<int>(std::round((m_range->firstRunTo - static_cast<int>(m_range->firstRunTo)) * 100)));
+  std::string bestRun = !isBestRunExists
+                            ? "<small>None</small>"
+                            : fmt::format(
+                                  "{}<small>.{:02d}%</small> - {}<small>.{:02d}%</small>",
+                                  static_cast<int>(m_range->bestRunFrom),
+                                  static_cast<int>(std::round((m_range->bestRunFrom - static_cast<int>(m_range->bestRunFrom)) * 100)),
+                                  static_cast<int>(m_range->bestRunTo),
+                                  static_cast<int>(std::round((m_range->bestRunTo - static_cast<int>(m_range->bestRunTo)) * 100)));
+
+  std::vector<MetaData> tableData = {
+      {"Attempts:", attempts},
+      {"Completions:", completions},
+      {"Best Run:", bestRun},
+      {"First Run:", firstRun},
+      // {"Time Played:", "5<small>h</small> 20<small>m</small> 32<small>s</small>"},
+      {"Time Played:", "beta"},
+  };
+
+  m_table = MetaTable::create(tableData, m_head->getContentWidth(), {0, 5.f, 5.f, 5.f});
+  m_table->setVisible(m_isExpanded);
+  m_content->addChild(m_table);
 
   return true;
 }
 
+void StageRangeCell::updateLayoutWrapper()
+{
+  m_content->updateLayout();
+  m_content->setPositionY(m_content->getContentHeight());
+  this->setContentSize(m_content->getContentSize());
+
+  updateBackgroundTexture();
+  UpdateScrollLayoutEvent().post();
+}
+
 void StageRangeCell::updateBackgroundTexture()
 {
-  const auto cellSize = this->getContentSize();
+  if (m_background)
+    m_background->removeFromParentAndCleanup(true);
+
+  const auto contentSize = m_content->getContentSize();
   const auto bg_spr = m_disabled
                           ? "range-disabled-bg.png"_spr
                       : m_checked   ? "range-completed-bg.png"_spr
@@ -82,11 +156,65 @@ void StageRangeCell::updateBackgroundTexture()
                                     : "range-default-bg.png"_spr;
 
   m_background = CCScale9Sprite::create(bg_spr);
-  m_background->setContentSize(cellSize);
-  m_background->setPosition({cellSize.width / 2, cellSize.height / 2});
+  m_background->setContentSize(contentSize);
+  m_background->setPosition({contentSize.width / 2, contentSize.height / 2});
   m_background->setZOrder(0);
 
   this->addChild(m_background);
+}
+
+void StageRangeCell::updateExpandButton()
+{
+  if (m_expandBtnMenu)
+    m_expandBtnMenu->removeFromParentAndCleanup(true);
+
+  const auto cellSize = m_head->getContentSize();
+
+  // ! ---  Expand Button Menu --- !
+  m_expandBtnMenu = CCMenu::create();
+  m_expandBtnMenu->setAnchorPoint({1.f, 0.5f});
+  m_expandBtnMenu->setPosition({cellSize.width - 5.f, cellSize.height / 2});
+  m_expandBtnMenu->setLayout(
+      RowLayout::create()
+          ->setGap(2.5f)
+          ->setAutoScale(false)
+          ->setAutoGrowAxis(true)
+          ->setAxisAlignment(AxisAlignment::End)
+          ->setCrossAxisAlignment(AxisAlignment::Center));
+  m_expandBtnMenu->getLayout()->ignoreInvisibleChildren(true);
+  m_head->addChild(m_expandBtnMenu);
+
+  // ! ---  Expand Button --- !
+  const auto sprName = m_isExpanded
+                           ? "purple-chevron-up-btn.png"_spr
+                           : "purple-chevron-down-btn.png"_spr;
+  auto spr = CCSprite::createWithSpriteFrameName(sprName);
+  const auto expandBtn = CCMenuItemSpriteExtra::create(spr, this, menu_selector(StageRangeCell::onExpand));
+  expandBtn->ignoreAnchorPointForPosition(true);
+  expandBtn->setScale(.5f);
+  expandBtn->m_baseScale = expandBtn->getScale();
+
+  m_expandBtnMenu->addChild(expandBtn);
+  m_expandBtnMenu->updateLayout();
+}
+
+void StageRangeCell::updateTextColors()
+{
+  if (m_rangeLabel)
+  {
+    if (m_checked)
+      m_rangeLabel->setVariant(Label::Variant::Green);
+    else if (m_disabled)
+      m_rangeLabel->setVariant(Label::Variant::Red);
+    else
+      m_rangeLabel->setVariant(Label::Variant::Orange);
+  }
+}
+
+void StageRangeCell::updateMetaContent()
+{
+  if (m_table)
+    m_table->setVisible(m_isExpanded);
 }
 
 void StageRangeCell::onToggle(CCObject *sender)
@@ -112,6 +240,10 @@ void StageRangeCell::onToggle(CCObject *sender)
       if (range.from == m_range->from && range.to == m_range->to)
       {
         range.checked = !range.checked;
+
+        if (range.checked && range.completionCounter <= 0)
+          range.completionCounter = 1;
+
         break;
       }
     }
@@ -127,17 +259,13 @@ void StageRangeCell::onToggle(CCObject *sender)
   }
 }
 
-void StageRangeCell::updateTextColors()
+void StageRangeCell::onExpand(CCObject *sender)
 {
-  if (m_rangeLabel)
-  {
-    if (m_checked)
-      m_rangeLabel->setVariant(Label::Variant::Green);
-    else if (m_disabled)
-      m_rangeLabel->setVariant(Label::Variant::Red);
-    else
-      m_rangeLabel->setVariant(Label::Variant::Orange);
-  }
+  m_isExpanded = !m_isExpanded;
+
+  updateExpandButton();
+  updateMetaContent();
+  updateLayoutWrapper();
 }
 
 void StageRangeCell::setDisabled(bool disabled)