+++ src/serialization/profile/Range.hpp
@@ -7,6 +7,10 @@ struct Range
   std::string id;
   float from;
   float to;
+  float firstRunFrom;
+  float firstRunTo;
+  float bestRunFrom;
+  float bestRunTo;
   bool checked;
   std::string note;
   int attempts;