+++ src/ui/Label.hpp
@@ -19,25 +19,28 @@ private:
 
   std::string m_font;
   std::vector<CCLabelBMFont *> m_parts;
+  std::string m_text;
+  float m_fontSize = .4f;
 
   const ccColor3B greenColor = {99, 224, 110};
   const ccColor3B orangeColor = {255, 196, 157};
   const ccColor3B redColor = {253, 106, 106};
 
   void updateColor();
-  void parseAndBuild(const std::string &text);
+  void parseAndBuild();
 
 public:
   static Label *create(const std::string &text,
-                       const std::string &font = "gjFont17.fnt");
-
+                       const std::string &font = "gjFont17.fnt",
+                       float fontSize = 1.f);
   bool init(const std::string &text,
-            const std::string &font = "gjFont17.fnt");
+            const std::string &font = "gjFont17.fnt",
+            float fontSize = 1.f);
 
   CCLabelBMFont *getLabel() const { return m_label; }
+  Variant getVariant() const;
 
   void setText(const std::string &text);
-
+  void setFontSize(float fontSize);
   void setVariant(Variant variant);
-  Variant getVariant() const;
 };